requirements.txt
boto3
botocore
flask
gunicorn
streamlit
crewai
crewai_tools
Pillow